requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
fitter==1.1.11
idna==2.8
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==3.0.3
nbformat==4.4.0
numpy==1.16.2
pandas==0.24.2
peewee==3.9.5
Pillow==6.0.0
plotly==3.9.0
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.3
